Move ForkAndSpecMilestone to spec.datastructures.util package
