+++ ethereum/spec/src/main/java/tech/pegasys/teku/spec/datastructures/util/ForkAndSpecMilestone.java
@@ -11,7 +11,7 @@
  * specific language governing permissions and limitations under the License.
  */
 
-package tech.pegasys.teku.spec.containers;
+package tech.pegasys.teku.spec.datastructures.util;
 
 import tech.pegasys.teku.spec.SpecMilestone;
 import tech.pegasys.teku.spec.datastructures.state.Fork;
